[Ref] Silence even more 64bit warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@10426 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/EffectVis.cpp
@@ -70,7 +70,7 @@ void CEffectVis::OnActionChanged()
 
 void CEffectVis::OnEffectChanged()
 {
-	m_nFillEffect = m_cmbEffectList.GetItemData(m_cmbEffectList.GetCurSel());
+	m_nFillEffect = static_cast<UINT>(m_cmbEffectList.GetItemData(m_cmbEffectList.GetCurSel()));
 }
 
 void CEffectVis::OnPaint()
